extract v2 data set importer
This extracted importer will be used by the DST plugin.
There is other code in the data-interactives folder which also handles converting v2 like datasets sent by plugins. However it wasn't easy to unify that code too.
